+++ src/main/java/com/example/portalegresso/model/repositorio/CargoRepositorio.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.portalegresso.model.entidades.Cargo;
 
-public interface CargoRepositorio extends JpaRepository<Cargo,Long> {
+public interface CargoRepositorio extends JpaRepository<Cargo,Integer> {
 
 }
