+++ src/main/java/com/example/portalegresso/model/repositorio/CursoEgressoRepositorio.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.portalegresso.model.entidades.CursoEgresso;
 
-public interface CursoEgressoRepositorio extends JpaRepository<CursoEgresso,Long> {
+public interface CursoEgressoRepositorio extends JpaRepository<CursoEgresso,Integer> {
 
 }
